requirements.txt
ccxt==4.1.83

aiogram==3.0.0

python-dotenv==1.0.1
